requirements.txt
fake-factory
ansible-tower-cli
pyyaml
requests